pyproject.toml
[project]
name = "spoolman-labels"
version = "0.1.0"
description = "Create label images for Spoolman spools"
readme = "README.md"
authors = [
    { name = "Laura Batalha", email = "5883822+lbatalha@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "brother-ql-inventree>=1.3",
    "pillow>=12.0.0",
    "qrcode[pil]>=8.2",
    "requests>=2.32.5",
]

[project.scripts]
spoolman-labels = "spoolman_labels:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
